+++ apps/api/src/routes/health/get.ts
@@ -1,0 +1,9 @@
+import express from "express";
+
+const router = express.Router();
+
+router.get("/", function (_request, response) {
+  response.json({ status: "ok" });
+});
+
+export default router;
